all ok till purcase with purchase entry with update , remning, update book detail and member tables

+++ src/main/java/com/raja/lib/acc/repository/LedgerRepository.java
@@ -2,5 +2,6 @@ package com.raja.lib.acc.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.raja.lib.acc.model.Ledger;
-public interface LedgerRepository extends JpaRepository<Ledger, Integer> {
+
+public interface LedgerRepository extends JpaRepository<Ledger, Long> {
 }
