+++ src/main/java/com/raja/lib/invt/repository/BookTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import com.raja.lib.invt.model.BookType;
 
 @Repository
-public interface BookTypeRepository extends JpaRepository<BookType, Integer> {
+public interface BookTypeRepository extends JpaRepository<BookType, Long> {
 }
