+++ src/main/java/com/raja/lib/invt/resposne/PurchaseDetailResponseDto.java
@@ -9,13 +9,13 @@ import lombok.Setter;
 @NoArgsConstructor
 public class PurchaseDetailResponseDto {
     private Long purchaseDetailId;
-    private Integer bookId;  
+    private Long bookId;  
     private String bookName;
     private int qty;
     private int rate;
     private int amount;
 
-    public PurchaseDetailResponseDto(Long purchaseDetailId, Integer bookId, String bookName, int qty, int rate, int amount) {
+    public PurchaseDetailResponseDto(Long purchaseDetailId, Long bookId, String bookName, int qty, int rate, int amount) {
         this.purchaseDetailId = purchaseDetailId;
         this.bookId = bookId;
         this.bookName = bookName;
